Move Dijkstra utilities to respective Graph classes
Instead of leaving them into the global scope, where they are also difficult to reach via autocompletion, we add them to Graph and WeightedGraph respectively.

I also renamed `pathDictToPath(from:to:pathDict:)` to a nicer `route(from:to:in:)`.

The methods continue to be non-generic because the use of the protocol type `Edge` in the `Graph` class does not allow them to use the `Edge` type itself; only those that conform to it.

`distanceArrayToVertexDict` was using a `WeightedGraph<T, W>` where `T` is not constrained to `Hashable`, but internally it was using `T` as a dictionary's `Key`, which needs to be `Hashable`.

I'll add conformance to `Hashable` in the next commit.