Refactor toposort() with an indices-based implementation
While the algorithm is unchanged, the new implementation relies on indices and a single enum (`Topomark`), rather than, potentially, value types that were being copied twice: in `sorted` and `tsNodes`.

Another change that adds a bit of overhead is checking the graph has edges at all; if it doesn't, an empty array is returned instead of the previous behavior that was still sorting the nodes. While that might not necessarily be the ideal behavior for `toposort()` itself, _it is_ necessary for `isDAG` which before the change was true for a graph with no edges.