Remove `mutating` workaround
The only reason I had employed the approach of using internal `inout` methods and extend the protocol `where Self: AnyObject` was because I wasn't able to use the refined `E` initializer in protocol that conformed to the Graph protocol.

It turns out that when the protocol is refined with `where E: WeightedEdge`, where `E: Edge` and `WeightedEdge: Edge`, one can initialize the refinement (`WeightedEdge`) by using the associatedtype `E`.

So instead of `WeightedEdge(source: source, target: target, weight: weight)` you'd use `E(source: source, target: target, weight: weight)`. It may seem straightforward, but using `WeightedEdge` leads to compiletime error “Protocol 'WeightedEdge' can only be used as a generic constraint because it has Self or associated type requirements.”

Finally, all methods should be implementable directly in protocols instead of going down to the types that conform to them, and this means the workaround is not necessary for the time being.

Should it be, I added a note at the end of the `mutating` extension in Graph.swift so that it's easy to revert the change. But until then, I prefer to remove unnecessary code that needed double maintenance.

This commit updates mutability within tests so that the change is reflected. Given that we are not reimplementing `mutating` methods for classes anymore, they are using the `mutating` ones and therefore need to be declared with `var` if they need to be mutated.