Don't call indexOfVertex(head) per neighbor.
Reduces the cycle detection complexity from O(V²) to O(V).

indexOfVertex(head) takes O(V) time, and if we call it `neighbors` times
(O(V)) then we end up with the O(V²) time complexity of the cycle
detection.

To fix this, we only need to move the head vertex index lookup to the
outside of the loop.